refactor: consolidate integration tests under integration/
Moves mysql-integration and sqlite-integration under integration/{mysql,sqlite}
to make room for future driver suites (e.g., postgres). Adds an INTEGRATION=1
gate, auto-configures the container runtime endpoint (Docker or Podman),
supports MariaDB via image selector, and adds a session-label sweep so
leaked containers get cleaned up when testcontainers drops a reference on
wait-strategy failure. Also adds ToMariaDB aliases over the MySQL renderers.
