+++ render.go
@@ -72,3 +72,21 @@ func ToParameterizedMySQL(in string, opts ...Opt) (s string, params []any, err e
 
 	return mysql.RenderParam(e)
 }
+
+// ToMariaDB renders a lucene expression as a MariaDB sql filter string.
+// It is a thin alias over ToMySQL: everything this package emits (backtick
+// quoting, `LIKE ... ESCAPE '#'`, `REGEXP`, `BETWEEN`, bool literals,
+// `?` placeholders) is identical on MySQL and MariaDB. The `^(...)$` regex
+// fallback avoids `(?:...)` so it works on every regex engine either
+// database has shipped (MySQL 5.7/MariaDB 10.4 Henry Spencer, MySQL 8.0 ICU,
+// MariaDB 10.5+ PCRE2).
+func ToMariaDB(in string, opts ...Opt) (string, error) {
+	return ToMySQL(in, opts...)
+}
+
+// ToParameterizedMariaDB renders a lucene expression as a MariaDB sql filter
+// string with ? placeholders. Alias over ToParameterizedMySQL; see ToMariaDB
+// for the compatibility rationale.
+func ToParameterizedMariaDB(in string, opts ...Opt) (s string, params []any, err error) {
+	return ToParameterizedMySQL(in, opts...)
+}
